Create component for generic table boody
